+++ src/pages/Footer/Footer.css
@@ -37,7 +37,7 @@
   }
 
   #footer-nav,
-  #footer-rules,
+  #footer-policies,
   #footer-links {
     list-style: none;
     margin: 0;
